refactor: move integration tests and fix links
