+++ src/modules/users/tests/integration/user.service.int-spec.ts
@@ -2,10 +2,10 @@ import { NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { Test } from "@nestjs/testing";
 import { randomUUID } from "crypto";
 import { UsersRole } from "@prisma/client";
-import { AppModule } from "../../../src/app.module";
-import { PrismaService } from "../../../src/config/prisma/prisma.service";
-import { UsersService } from "../../../src/modules/users/users.service";
-import { CreateUserDto } from "../../../src/modules/users/dto";
+import { AppModule } from "../../../../app.module";
+import { PrismaService } from "../../../../config/prisma/prisma.service";
+import { UsersService } from "../../../users/users.service";
+import { CreateUserDto } from "../../../users/dto";
 
 describe("UsersService Integraton", () => {
   let prisma: PrismaService;
